requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
numpy==1.26.4
Pillow==10.4.0
python-pptx==1.0.2
python-multipart==0.0.20
